feat : fix get all invoice

+++ internal/model/invoice_response.go
@@ -1,8 +1,8 @@
 package model
 
 type InvoiceAggregateResponse struct {
-	TotalProfit            int `json:"total_profit"`
-	TotalOfCashTransaction int `json:"total_of_cash_transaction"`
+	TotalProfit            int64 `json:"total_profit"`
+	TotalOfCashTransaction int64 `json:"total_of_cash_transaction"`
 }
 
 type GetInvoiceResponse struct {
